refactor: split MessageHandling helpers out of 1908-line extension (#21846)
