+++ clients/shared/Features/Chat/ChatViewModel+MessageHandling.swift
@@ -70,386 +70,6 @@ extension ChatViewModel {
         }
     }
 
-
-    /// Substrings that indicate a tool failed because the OS denied permission.
-    /// This lets the UI reconcile "allowed" confirmations that still fail at
-    /// execution time (for example: user clicked Always Allow, then denied the
-    /// macOS Accessibility prompt).
-    private static let osPermissionDeniedIndicators: [String] = [
-        "accessibility permission not granted",
-        "accessibility permission denied",
-        "screen recording permission denied",
-        "full disk access",
-        "operation not permitted",
-        "permission denied",
-        "not authorized"
-    ]
-
-    /// Extract the most relevant tool input value as a full string (no truncation).
-    /// Redacts values for sensitive keys to prevent credential leakage into inputSummary.
-    func extractToolInput(_ input: [String: AnyCodable]) -> String {
-        HistoryReconstructionService.extractToolInputStatic(input)
-    }
-
-    /// Summarize tool input for display, picking the most relevant value truncated to 80 chars.
-    func summarizeToolInput(_ input: [String: AnyCodable]) -> String {
-        HistoryReconstructionService.summarizeToolInputStatic(input)
-    }
-
-    /// Format all tool input arguments for display in expanded details.
-    /// Primary key is listed first, then remaining keys alphabetically. All as `key: value`.
-    /// Sensitive keys (passwords, tokens, etc.) are redacted to prevent credential exposure.
-    func formatAllToolInput(_ input: [String: AnyCodable]) -> String {
-        guard !input.isEmpty else { return "" }
-
-        // Find the primary key (same logic as extractToolInput)
-        let primaryKey = HistoryReconstructionService.toolInputPriorityKeys.first(where: { input[$0] != nil })
-            ?? input.keys.sorted().first
-
-        // All keys as "key: value", primary key first then rest alphabetically
-        let orderedKeys: [String]
-        if let pk = primaryKey {
-            orderedKeys = [pk] + input.keys.filter { $0 != pk }.sorted()
-        } else {
-            orderedKeys = input.keys.sorted()
-        }
-
-        var lines: [String] = []
-        for key in orderedKeys {
-            guard let value = input[key] else { continue }
-            if HistoryReconstructionService.isSensitiveKey(key) {
-                lines.append("\(key): [redacted]")
-            } else {
-                lines.append("\(key): \(redactingStringifyValue(value))")
-            }
-        }
-
-        return lines.joined(separator: "\n")
-    }
-
-    private func stringifyValue(_ value: AnyCodable) -> String {
-        if let s = value.value as? String { return s }
-        if let b = value.value as? Bool { return b ? "true" : "false" }
-        if let n = value.value as? Int { return String(n) }
-        if let n = value.value as? Double { return String(n) }
-        if let encoder = try? JSONEncoder().encode(value),
-           let json = String(data: encoder, encoding: .utf8) {
-            return json
-        }
-        return String(describing: value.value ?? "")
-    }
-
-    /// Stringify a value, recursively redacting sensitive keys in nested objects.
-    private func redactingStringifyValue(_ value: AnyCodable) -> String {
-        if let dict = value.value as? [String: Any] {
-            return redactDictionary(dict)
-        }
-        if let array = value.value as? [Any] {
-            return redactArray(array)
-        }
-        return stringifyValue(value)
-    }
-
-    /// Recursively redact sensitive keys in a dictionary, returning a JSON-like string.
-    private func redactDictionary(_ dict: [String: Any]) -> String {
-        var redacted: [String: Any] = [:]
-        for (key, val) in dict {
-            if HistoryReconstructionService.isSensitiveKey(key) {
-                redacted[key] = "[redacted]"
-            } else if let nested = val as? [String: Any] {
-                redacted[key] = redactDictionaryAsObject(nested)
-            } else if let nested = val as? [Any] {
-                redacted[key] = redactArrayAsObject(nested)
-            } else {
-                redacted[key] = val
-            }
-        }
-        // Encode the redacted dict to JSON
-        if let data = try? JSONSerialization.data(withJSONObject: redacted, options: [.sortedKeys]),
-           let json = String(data: data, encoding: .utf8) {
-            return json
-        }
-        return String(describing: redacted)
-    }
-
-    /// Recursively redact sensitive keys in array elements.
-    private func redactArray(_ array: [Any]) -> String {
-        let redacted = redactArrayAsObject(array)
-        if let data = try? JSONSerialization.data(withJSONObject: redacted, options: [.sortedKeys]),
-           let json = String(data: data, encoding: .utf8) {
-            return json
-        }
-        return String(describing: redacted)
-    }
-
-    /// Recursively redact sensitive keys in array elements, returning an array (not string) for nesting.
-    private func redactArrayAsObject(_ array: [Any]) -> [Any] {
-        return array.map { element -> Any in
-            if let dict = element as? [String: Any] {
-                return redactDictionaryAsObject(dict)
-            } else if let nested = element as? [Any] {
-                return redactArrayAsObject(nested)
-            }
-            return element
-        }
-    }
-
-    /// Recursively redact sensitive keys, returning a dictionary (not string) for nesting.
-    private func redactDictionaryAsObject(_ dict: [String: Any]) -> [String: Any] {
-        var redacted: [String: Any] = [:]
-        for (key, val) in dict {
-            if HistoryReconstructionService.isSensitiveKey(key) {
-                redacted[key] = "[redacted]"
-            } else if let nested = val as? [String: Any] {
-                redacted[key] = redactDictionaryAsObject(nested)
-            } else if let nested = val as? [Any] {
-                redacted[key] = redactArrayAsObject(nested)
-            } else {
-                redacted[key] = val
-            }
-        }
-        return redacted
-    }
-
-    func toolDisplayName(_ name: String) -> String {
-        switch name {
-        case "file_write": return "Write File"
-        case "file_edit": return "Edit File"
-        case "bash": return "Run Command"
-        case "web_fetch": return "Fetch URL"
-        case "file_read": return "Read File"
-        case "glob": return "Find Files"
-        case "grep": return "Search Files"
-        default: return name.replacingOccurrences(of: "_", with: " ").capitalized
-        }
-    }
-
-    /// Extract a code preview from accumulated tool input JSON.
-    /// Shows the HTML code as it streams during app_create/app_refresh/app_update.
-    static func extractCodePreview(from accumulatedJson: String, toolName: String) -> String? {
-        guard !accumulatedJson.isEmpty else { return nil }
-        let isAppTool = toolName == "app_create" || toolName == "app_refresh" || toolName == "app_update"
-        guard isAppTool else { return nil }
-
-        // Find the html JSON string value by locating the opening quote
-        let markers = ["\"html\": \"", "\"html\":\""]
-        for marker in markers {
-            guard let range = accumulatedJson.range(of: marker) else { continue }
-            let afterMarker = accumulatedJson[range.upperBound...]
-
-            // Scan for the closing unescaped quote of the JSON string value
-            var result: [Character] = []
-            var i = afterMarker.startIndex
-            while i < afterMarker.endIndex {
-                let ch = afterMarker[i]
-                if ch == "\\" {
-                    let next = afterMarker.index(after: i)
-                    if next < afterMarker.endIndex {
-                        // Single-pass unescape: handle the pair
-                        switch afterMarker[next] {
-                        case "n": result.append("\n")
-                        case "t": result.append("\t")
-                        case "\"": result.append("\"")
-                        case "\\": result.append("\\")
-                        default:
-                            result.append(ch)
-                            result.append(afterMarker[next])
-                        }
-                        i = afterMarker.index(after: next)
-                    } else {
-                        // Trailing backslash (incomplete escape at end of stream)
-                        break
-                    }
-                } else if ch == "\"" {
-                    // Found the closing quote — stop
-                    break
-                } else {
-                    result.append(ch)
-                    i = afterMarker.index(after: i)
-                }
-            }
-
-            let html = String(result)
-            return html.isEmpty ? nil : html
-        }
-
-        return nil
-    }
-
-    /// Map attachment DTOs to ChatAttachment values, generating thumbnails for images.
-    func mapMessageAttachments(_ attachments: [UserMessageAttachment]) -> [ChatAttachment] {
-        HistoryReconstructionService.mapMessageAttachmentsStatic(attachments)
-    }
-
-    /// Ingest attachments from a completion/handoff event into the current or new assistant message.
-    func ingestAssistantAttachments(_ attachments: [UserMessageAttachment]?) {
-        guard let attachments, !attachments.isEmpty else { return }
-        let chatAttachments = mapMessageAttachments(attachments)
-        guard !chatAttachments.isEmpty else { return }
-
-        if let existingId = currentAssistantMessageId,
-           let index = messages.firstIndex(where: { $0.id == existingId }) {
-            messages[index].attachments.append(contentsOf: chatAttachments)
-        } else {
-            let msg = ChatMessage(role: .assistant, text: "", attachments: chatAttachments)
-            currentAssistantMessageId = msg.id
-            messages.append(msg)
-        }
-    }
-
-    /// Ingest attachment warnings from a completion/handoff event into the
-    /// current or new assistant message.
-    func ingestAssistantAttachmentWarnings(_ warnings: [String]?) {
-        guard let warnings, !warnings.isEmpty else { return }
-
-        if let existingId = currentAssistantMessageId,
-           let index = messages.firstIndex(where: { $0.id == existingId }) {
-            messages[index].attachmentWarnings.append(contentsOf: warnings)
-        } else {
-            let msg = ChatMessage(role: .assistant, text: "", attachmentWarnings: warnings)
-            currentAssistantMessageId = msg.id
-            messages.append(msg)
-        }
-    }
-
-    /// Returns true when a tool error string looks like a macOS/TCC permission denial.
-    private static func isOSPermissionDeniedError(_ result: String) -> Bool {
-        let normalized = result.lowercased()
-        return osPermissionDeniedIndicators.contains { normalized.contains($0) }
-    }
-
-    /// If the user approved a confirmation but execution still failed due OS
-    /// permission denial, update the nearby confirmation so the UI does not
-    /// incorrectly show it as approved.
-    private func downgradeAdjacentApprovedConfirmationForPermissionDeniedError(
-        assistantMessageIndex: Int,
-        toolResult: String,
-        isError: Bool
-    ) {
-        guard isError, Self.isOSPermissionDeniedError(toolResult) else { return }
-
-        var index = assistantMessageIndex + 1
-        while index < messages.count {
-            // Stay within this turn.
-            if messages[index].role == .user { break }
-
-            guard messages[index].confirmation != nil else {
-                index += 1
-                continue
-            }
-
-            if messages[index].confirmation?.state == .approved {
-                messages[index].confirmation?.state = .denied
-            }
-            return
-        }
-    }
-
-    // MARK: - Streaming Delta Throttle
-
-    /// Cancel any pending flush and discard buffered text.
-    /// Called on every path that clears `currentAssistantMessageId` without
-    /// a normal `messageComplete` (cancel, error, handoff, reconnect, etc.)
-    /// to prevent a stale flush from creating an orphan assistant message.
-    func discardStreamingBuffer() {
-        streamingFlushTask?.cancel()
-        streamingFlushTask = nil
-        streamingDeltaBuffer = ""
-    }
-
-    /// Flush any buffered streaming text into the messages array.
-    /// Called on a timer and also eagerly on `messageComplete`,
-    /// `toolUseStart`, `uiSurfaceShow`, etc.
-    func flushStreamingBuffer() {
-        streamingFlushTask?.cancel()
-        streamingFlushTask = nil
-        guard !streamingDeltaBuffer.isEmpty else { return }
-        let buffered = streamingDeltaBuffer
-        streamingDeltaBuffer = ""
-        appendTextToCurrentMessage(buffered)
-    }
-
-    /// Append a chunk of text to the current assistant message.
-    private func appendTextToCurrentMessage(_ text: String) {
-        guard !text.isEmpty else { return }
-        if let existingId = currentAssistantMessageId,
-           let index = messages.firstIndex(where: { $0.id == existingId }) {
-            if lastContentWasToolCall || messages[index].textSegments.isEmpty {
-                let segIdx = messages[index].textSegments.count
-                messages[index].textSegments.append(text)
-                messages[index].contentOrder.append(.text(segIdx))
-                lastContentWasToolCall = false
-            } else {
-                messages[index].textSegments[messages[index].textSegments.count - 1] += text
-            }
-        } else if currentAssistantMessageId != nil {
-            log.warning("Stale currentAssistantMessageId \(self.currentAssistantMessageId!.uuidString) — discarding \(text.count) buffered chars")
-            currentAssistantMessageId = nil
-            return
-        } else {
-            var msg = ChatMessage(role: .assistant, text: text, isStreaming: true)
-            if currentTurnUserText == "/models" {
-                msg.modelList = ModelListData()
-            } else if currentTurnUserText == "/commands" {
-                msg.commandList = CommandListData()
-            }
-            currentAssistantMessageId = msg.id
-            messages.append(msg)
-            lastContentWasToolCall = false
-        }
-    }
-
-    /// Schedule a flush after the throttle interval if one isn't already pending.
-    private func scheduleStreamingFlush() {
-        guard streamingFlushTask == nil else { return }
-        streamingFlushTask = Task { @MainActor [weak self] in
-            try? await Task.sleep(nanoseconds: UInt64(Self.streamingFlushInterval * 1_000_000_000))
-            guard !Task.isCancelled else { return }
-            self?.flushStreamingBuffer()
-        }
-    }
-
-    // MARK: - Partial Output Coalescing
-
-    /// Flush buffered partial-output chunks into the `messages` array.
-    /// Called on a timer and eagerly on `messageComplete` / `toolResult`.
-    func flushPartialOutputBuffer() {
-        partialOutputFlushTask?.cancel()
-        partialOutputFlushTask = nil
-        guard !partialOutputBuffer.isEmpty else { return }
-        let buffered = partialOutputBuffer
-        partialOutputBuffer = [:]
-        let maxPartialOutput = 5000
-        for (_, entry) in buffered {
-            let tcIndex = entry.tcIndex
-            guard let msgIndex = messages.firstIndex(where: { $0.id == entry.messageId }),
-                  tcIndex < messages[msgIndex].toolCalls.count else { continue }
-            messages[msgIndex].toolCalls[tcIndex].partialOutput.append(entry.content)
-            if messages[msgIndex].toolCalls[tcIndex].partialOutput.count > maxPartialOutput {
-                let excess = messages[msgIndex].toolCalls[tcIndex].partialOutput.count - maxPartialOutput
-                messages[msgIndex].toolCalls[tcIndex].partialOutput.removeFirst(excess)
-            }
-            messages[msgIndex].toolCalls[tcIndex].partialOutputRevision += 1
-        }
-    }
-
-    /// Discard any buffered partial-output chunks without flushing.
-    func discardPartialOutputBuffer() {
-        partialOutputFlushTask?.cancel()
-        partialOutputFlushTask = nil
-        partialOutputBuffer = [:]
-    }
-
-    /// Schedule a partial-output flush after the throttle interval if one isn't already pending.
-    private func schedulePartialOutputFlush() {
-        guard partialOutputFlushTask == nil else { return }
-        partialOutputFlushTask = Task { @MainActor [weak self] in
-            try? await Task.sleep(nanoseconds: UInt64(Self.streamingFlushInterval * 1_000_000_000))
-            guard !Task.isCancelled else { return }
-            self?.flushPartialOutputBuffer()
-        }
-    }
-
     public func handleServerMessage(_ message: ServerMessage) {
         switch message {
         case .conversationInfo(let info):
@@ -1141,437 +761,34 @@ extension ChatViewModel {
             }
 
         case .toolUsePreviewStart(let msg):
-            guard belongsToConversation(msg.conversationId) else { return }
-            guard !isCancelling else { return }
-            guard !isLoadingHistory else { return }
-            guard !isWorkspaceRefinementInFlight else { return }
-            // Suppress preview chip for proxy tools — the inline surface widget replaces them.
-            if msg.toolName == "ui_show" || msg.toolName == "ui_update" || msg.toolName == "ui_dismiss" {
-                break
-            }
-            // Flush buffered text so it lands before the tool call in content order.
-            flushStreamingBuffer()
-            flushPartialOutputBuffer()
-            // If a chip with the same toolUseId already exists (e.g. toolUseStart
-            // arrived before this preview), ignore the late preview.
-            if let existingId = currentAssistantMessageId,
-               let index = messages.firstIndex(where: { $0.id == existingId }),
-               messages[index].toolCalls.contains(where: { $0.toolUseId == msg.toolUseId }) {
-                break
-            }
-            isThinking = false
-            var toolCall = ToolCallData(
-                toolName: msg.toolName,
-                inputSummary: "Preparing...",
-                inputFull: "",
-                inputRawValue: "",
-                arrivedBeforeText: !currentAssistantHasText,
-                startedAt: Date()
-            )
-            toolCall.toolUseId = msg.toolUseId
-            // Add to existing assistant message or create one.
-            if let existingId = currentAssistantMessageId,
-               let index = messages.firstIndex(where: { $0.id == existingId }),
-               messages[index].toolCalls.count < 100 {
-                let tcIdx = messages[index].toolCalls.count
-                messages[index].toolCalls.append(toolCall)
-                messages[index].contentOrder.append(.toolCall(tcIdx))
-            } else {
-                if let existingId = currentAssistantMessageId,
-                   let oldIndex = messages.firstIndex(where: { $0.id == existingId }) {
-                    messages[oldIndex].isStreaming = false
-                }
-                var newMsg = ChatMessage(role: .assistant, text: "", isStreaming: true, toolCalls: [toolCall])
-                newMsg.contentOrder = [.toolCall(0)]
-                currentAssistantMessageId = newMsg.id
-                messages.append(newMsg)
-            }
-            lastContentWasToolCall = true
+            handleToolUsePreviewStart(msg)
 
         case .toolUseStart(let msg):
-            guard belongsToConversation(msg.conversationId) else { return }
-            guard !isCancelling else { return }
-            guard !isLoadingHistory else { return }
-            guard !isWorkspaceRefinementInFlight else { return }
-            // Flush buffered text so it lands before the tool call in content order.
-            flushStreamingBuffer()
-            flushPartialOutputBuffer()
-            lastToolUseReceivedAt = Date()
-            // Suppress ToolCallChip for ui_show — the inline surface widget replaces it.
-            if msg.toolName == "ui_show" || msg.toolName == "ui_update" || msg.toolName == "ui_dismiss" {
-                break
-            }
-            // Tool chip is now visible — hide the thinking indicator
-            isThinking = false
-            // Extract building status for app tools
-            let buildingStatus: String? = {
-                let appTools: Set<String> = ["app_create", "app_refresh", "app_update"]
-                guard appTools.contains(msg.toolName) else { return nil }
-                if let status = msg.input["status"]?.value as? String, !status.isEmpty {
-                    return status
-                }
-                // app_create/app_refresh/app_update rely on friendlyRunningLabel + progressive label cycling
-                return nil
-            }()
-            // Upsert by toolUseId: if a preview chip already exists for this tool, update it
-            // instead of creating a duplicate.
-            if let toolUseId = msg.toolUseId,
-               let existingId = currentAssistantMessageId,
-               let msgIndex = messages.firstIndex(where: { $0.id == existingId }),
-               let tcIndex = messages[msgIndex].toolCalls.firstIndex(where: { $0.toolUseId == toolUseId }) {
-                messages[msgIndex].toolCalls[tcIndex].inputSummary = summarizeToolInput(msg.input)
-                messages[msgIndex].toolCalls[tcIndex].inputFull = formatAllToolInput(msg.input)
-                messages[msgIndex].toolCalls[tcIndex].inputRawValue = extractToolInput(msg.input)
-                messages[msgIndex].toolCalls[tcIndex].inputRawDict = msg.input
-                messages[msgIndex].toolCalls[tcIndex].buildingStatus = buildingStatus
-                messages[msgIndex].toolCalls[tcIndex].reasonDescription = (msg.input["activity"]?.value as? String)
-                    ?? (msg.input["reason"]?.value as? String)
-                    ?? (msg.input["reasoning"]?.value as? String)
-                break
-            }
-            var toolCall = ToolCallData(
-                toolName: msg.toolName,
-                inputSummary: summarizeToolInput(msg.input),
-                inputFull: formatAllToolInput(msg.input),
-                inputRawValue: extractToolInput(msg.input),
-                arrivedBeforeText: !currentAssistantHasText,
-                startedAt: Date()
-            )
-            toolCall.buildingStatus = buildingStatus
-            toolCall.toolUseId = msg.toolUseId
-            toolCall.inputRawDict = msg.input
-            toolCall.reasonDescription = (msg.input["activity"]?.value as? String)
-                ?? (msg.input["reason"]?.value as? String)
-                ?? (msg.input["reasoning"]?.value as? String)
-            // Add to existing assistant message or create one.
-            // Cap at 100 tool calls per message to prevent unbounded memory growth;
-            // overflow falls through to create a new message.
-            if let existingId = currentAssistantMessageId,
-               let index = messages.firstIndex(where: { $0.id == existingId }),
-               messages[index].toolCalls.count < 100 {
-                let tcIdx = messages[index].toolCalls.count
-                messages[index].toolCalls.append(toolCall)
-                messages[index].contentOrder.append(.toolCall(tcIdx))
-            } else {
-                // Cap reached — rotate to a new message.
-                // Clear streaming state on the old message first.
-                if let existingId = currentAssistantMessageId,
-                   let oldIndex = messages.firstIndex(where: { $0.id == existingId }) {
-                    messages[oldIndex].isStreaming = false
-                }
-                var newMsg = ChatMessage(role: .assistant, text: "", isStreaming: true, toolCalls: [toolCall])
-                newMsg.contentOrder = [.toolCall(0)]
-                currentAssistantMessageId = newMsg.id
-                messages.append(newMsg)
-            }
-            lastContentWasToolCall = true
+            handleToolUseStart(msg)
 
         case .toolInputDelta(let msg):
-            guard belongsToConversation(msg.conversationId) else { return }
-            guard !isCancelling else { return }
-            guard !isLoadingHistory else { return }
-            let preview = Self.extractCodePreview(from: msg.content, toolName: msg.toolName)
-            // If toolUseId is present, find the matching tool call and update its streaming preview.
-            if let toolUseId = msg.toolUseId,
-               let existingId = currentAssistantMessageId,
-               let msgIndex = messages.firstIndex(where: { $0.id == existingId }),
-               let tcIndex = messages[msgIndex].toolCalls.firstIndex(where: { $0.toolUseId == toolUseId }) {
-                // Update code preview on the message for the matched tool call
-                messages[msgIndex].streamingCodePreview = preview
-                messages[msgIndex].streamingCodeToolName = msg.toolName
-                _ = tcIndex // suppress unused warning — match confirms the tool call exists
-            } else if let existingId = currentAssistantMessageId,
-               let msgIndex = messages.firstIndex(where: { $0.id == existingId }) {
-                messages[msgIndex].streamingCodePreview = preview
-                messages[msgIndex].streamingCodeToolName = msg.toolName
-            } else {
-                var newMsg = ChatMessage(role: .assistant, text: "", isStreaming: true)
-                newMsg.streamingCodePreview = preview
-                newMsg.streamingCodeToolName = msg.toolName
-                currentAssistantMessageId = newMsg.id
-                messages.append(newMsg)
-            }
+            handleToolInputDelta(msg)
 
         case .toolOutputChunk(let msg):
-            guard !isCancelling else { return }
-            guard belongsToConversation(msg.conversationId) else { return }
-            guard !isLoadingHistory else { return }
-            if (msg.subType == nil || msg.subType?.isEmpty == true),
-               !msg.chunk.isEmpty {
-                // Resolve target tool call: prefer toolUseId, fall back to positional heuristic.
-                let resolvedPlainTarget: (msgIndex: Int, tcIndex: Int)? = {
-                    if let toolUseId = msg.toolUseId {
-                        for i in stride(from: messages.count - 1, through: 0, by: -1) {
-                            if let tcIdx = messages[i].toolCalls.firstIndex(where: { $0.toolUseId == toolUseId }) {
-                                return (i, tcIdx)
-                            }
-                        }
-                    }
-                    if let existingId = currentAssistantMessageId,
-                       let mIdx = messages.firstIndex(where: { $0.id == existingId }),
-                       let tcIdx = messages[mIdx].toolCalls.lastIndex(where: { !$0.isComplete }) {
-                        return (mIdx, tcIdx)
-                    }
-                    return nil
-                }()
-                guard let target = resolvedPlainTarget else { return }
-                let msgIndex = target.msgIndex
-                let tcIndex = target.tcIndex
-                // Append plain-text output chunks to the coalescing buffer.
-                // Structured JSON sub-events (with a valid subType) are handled above;
-                // the subType guard prevents them from leaking raw JSON here.
-                let messageId = messages[msgIndex].id
-                let key = "\(messageId):\(tcIndex)"
-                if var entry = partialOutputBuffer[key] {
-                    entry.content += msg.chunk
-                    partialOutputBuffer[key] = entry
-                } else {
-                    partialOutputBuffer[key] = (messageId: messageId, tcIndex: tcIndex, content: msg.chunk)
-                }
-                schedulePartialOutputFlush()
-            }
+            handleToolOutputChunk(msg)
 
         case .toolResult(let msg):
-            guard belongsToConversation(msg.conversationId) else { return }
-            guard !isCancelling else { return }
-            guard !isLoadingHistory else { return }
-            guard !isWorkspaceRefinementInFlight else { return }
-            flushPartialOutputBuffer()
-            // Find the matching tool call.
-            // Prefer matching by toolUseId (stable identifier) over positional heuristics.
-            var targetMsgIndex: Int?
-            var targetTcIndex: Int?
-            if let toolUseId = msg.toolUseId {
-                // Search all messages for a tool call with matching toolUseId
-                for i in stride(from: messages.count - 1, through: 0, by: -1) {
-                    if let tcIndex = messages[i].toolCalls.firstIndex(where: { $0.toolUseId == toolUseId }) {
-                        targetMsgIndex = i
-                        targetTcIndex = tcIndex
-                        break
-                    }
-                }
-            }
-            // Fall back to existing positional heuristic if no ID match.
-            if targetMsgIndex == nil {
-                if let existingId = currentAssistantMessageId,
-                   let msgIndex = messages.firstIndex(where: { $0.id == existingId }),
-                   let tcIndex = messages[msgIndex].toolCalls.lastIndex(where: { !$0.isComplete }) {
-                    targetMsgIndex = msgIndex
-                    targetTcIndex = tcIndex
-                } else if let existingId = currentAssistantMessageId,
-                          let currentIdx = messages.firstIndex(where: { $0.id == existingId }) {
-                    // Current assistant message has no incomplete tool calls.
-                    // Search backward from current message position for rotated messages.
-                    for i in stride(from: currentIdx - 1, through: max(0, currentIdx - 5), by: -1) {
-                        guard messages[i].role == .assistant else { continue }
-                        if let tcIndex = messages[i].toolCalls.lastIndex(where: { !$0.isComplete }) {
-                            targetMsgIndex = i
-                            targetTcIndex = tcIndex
-                            break
-                        }
-                    }
-                } else {
-                    // currentAssistantMessageId is nil — search backward within current turn
-                    // (reconnect scenario where there are no queued messages).
-                    let lastUserIndex = messages.lastIndex(where: { $0.role == .user }) ?? 0
-                    for i in stride(from: messages.count - 1, through: lastUserIndex, by: -1) {
-                        guard messages[i].role == .assistant else { continue }
-                        if let tcIndex = messages[i].toolCalls.lastIndex(where: { !$0.isComplete }) {
-                            targetMsgIndex = i
-                            targetTcIndex = tcIndex
-                            break
-                        }
-                    }
-                }
-            }
-            if let msgIndex = targetMsgIndex, let tcIndex = targetTcIndex {
-                messages[msgIndex].toolCalls[tcIndex].result = msg.result
-                messages[msgIndex].toolCalls[tcIndex].resultLength = msg.result.count
-                messages[msgIndex].toolCalls[tcIndex].isError = msg.isError ?? false
-                messages[msgIndex].toolCalls[tcIndex].isComplete = true
-                messages[msgIndex].toolCalls[tcIndex].completedAt = Date()
-                let decoded = ToolCallData.decodeImage(from: msg.imageData)
-                // Keep cachedImage for display, nil out raw base64 to save ~2.7MB per screenshot
-                messages[msgIndex].toolCalls[tcIndex].cachedImage = decoded
-                messages[msgIndex].toolCalls[tcIndex].imageData = decoded == nil ? msg.imageData : nil
-                if let status = msg.status, !status.isEmpty {
-                    messages[msgIndex].toolCalls[tcIndex].buildingStatus = status
-                }
-                let toolErrored = msg.isError ?? false
-                downgradeAdjacentApprovedConfirmationForPermissionDeniedError(
-                    assistantMessageIndex: msgIndex,
-                    toolResult: msg.result,
-                    isError: toolErrored
-                )
-                if toolErrored, Self.isOSPermissionDeniedError(msg.result),
-                   messages[msgIndex].toolCalls[tcIndex].confirmationDecision == .approved {
-                    messages[msgIndex].toolCalls[tcIndex].confirmationDecision = .denied
-                }
-            }
-            // Auto-open clip files in the default video player.
-            // Use msg.toolName from the event payload (stable) instead of the
-            // matched tool call's toolName (relies on last-incomplete heuristic).
-            autoOpenClipIfNeeded(
-                toolName: msg.toolName,
-                result: msg.result,
-                isError: msg.isError ?? false
-            )
-
-            // Tool completed — don't re-show "Thinking..." here. The tool
-            // call chip already indicates activity, and the LLM isn't actually
-            // thinking yet. isThinking will be set when the user sends a new
-            // message or the daemon echoes it back.
+            handleToolResult(msg)
 
         case .uiSurfaceShow(let msg):
-            guard belongsToConversation(msg.conversationId) else { return }
-            guard msg.display == nil || msg.display == "inline" || msg.display == "panel" else { break }
-            guard let surface = Surface.from(msg) else { break }
-
-            // Flush buffered text so it lands before the surface in content order.
-            flushStreamingBuffer()
-            flushPartialOutputBuffer()
-
-            // Show floating overlay for task_progress cards (macOS only)
-            #if os(macOS)
-            if case .card(let cardData) = surface.data,
-               cardData.template == "task_progress",
-               let templateData = cardData.templateData,
-               let progressData = TaskProgressData.parse(from: templateData, fallbackTitle: cardData.title) {
-                TaskProgressOverlayManager.shared.show(data: progressData, surfaceId: msg.surfaceId)
-            }
-            #endif
-
-            // On macOS, dynamic pages with no explicit display mode (or "panel")
-            // are routed to the workspace by SurfaceManager. If the dynamic page
-            // has a preview, also render a compact preview card inline in chat.
-            // On iOS there is no workspace, so dynamic pages always render inline.
-            #if os(macOS)
-            if case .dynamicPage(let dpData) = surface.data, msg.display == nil || msg.display == "panel" {
-                isThinking = false
-                // Only render inline preview if the dynamic page has preview metadata
-                guard dpData.preview != nil else {
-                    log.info("Skipping inline surface - no preview metadata")
-                    break
-                }
-            } else if msg.display == "panel" {
-                // Non-dynamic-page surfaces with "panel" display are rendered as
-                // floating panels by SurfaceManager — skip inline rendering to
-                // avoid showing duplicates (one inline, one in a panel window).
-                break
-            }
-            #endif
-
-            isThinking = false
-            let inlineSurface = InlineSurfaceData(
-                id: surface.id,
-                surfaceType: surface.type,
-                title: surface.title,
-                data: surface.data,
-                actions: surface.actions,
-                surfaceRef: SurfaceRef(from: msg, surface: surface)
-            )
-
-            if let existingId = currentAssistantMessageId,
-               let index = messages.firstIndex(where: { $0.id == existingId }) {
-                let surfIdx = messages[index].inlineSurfaces.count
-                messages[index].inlineSurfaces.append(inlineSurface)
-                messages[index].contentOrder.append(.surface(surfIdx))
-                lastContentWasToolCall = true
-            } else if let lastUserIndex = messages.lastIndex(where: { $0.role == .user }),
-                      let idx = messages[lastUserIndex...].lastIndex(where: { $0.role == .assistant }) {
-                let surfIdx = messages[idx].inlineSurfaces.count
-                messages[idx].inlineSurfaces.append(inlineSurface)
-                messages[idx].contentOrder.append(.surface(surfIdx))
-                lastContentWasToolCall = true
-            } else {
-                var newMsg = ChatMessage(role: .assistant, text: "", isStreaming: true, inlineSurfaces: [inlineSurface])
-                newMsg.contentOrder = [.surface(0)]
-                currentAssistantMessageId = newMsg.id
-                messages.append(newMsg)
-            }
-
-            // Eagerly request preview for app surfaces that don't have one yet.
-            // Include the HTML so the handler can fall back to offscreen capture
-            // when the daemon has no stored preview (e.g. brand new app).
-            if case .dynamicPage(let dpData) = surface.data,
-               let appId = dpData.appId,
-               dpData.preview != nil,
-               dpData.preview?.previewImage == nil {
-                var userInfo: [String: Any] = ["appId": appId]
-                userInfo["html"] = dpData.html
-                NotificationCenter.default.post(
-                    name: Notification.Name("MainWindow.requestAppPreview"),
-                    object: nil,
-                    userInfo: userInfo
-                )
-            }
+            handleSurfaceShow(msg)
 
         case .uiSurfaceUndoResult(let msg):
-            guard belongsToConversation(msg.conversationId) else { return }
-            surfaceUndoCount = msg.remainingUndos
+            handleSurfaceUndoResult(msg)
 
         case .uiSurfaceUpdate(let msg):
-            guard belongsToConversation(msg.conversationId) else { return }
-            if isWorkspaceRefinementInFlight {
-                refinementReceivedSurfaceUpdate = true
-            }
-            if msg.surfaceId == activeSurfaceId {
-                surfaceUndoCount += 1
-            }
-            // Find the inline surface across all messages and update its data
-            for msgIndex in messages.indices {
-                if let surfaceIndex = messages[msgIndex].inlineSurfaces.firstIndex(where: { $0.id == msg.surfaceId }) {
-                    let existing = messages[msgIndex].inlineSurfaces[surfaceIndex]
-                    let tempSurface = Surface(id: existing.id, conversationId: msg.conversationId, type: existing.surfaceType, title: existing.title, data: existing.data, actions: existing.actions)
-                    if let updated = tempSurface.updated(with: msg) {
-                        messages[msgIndex].inlineSurfaces[surfaceIndex] = InlineSurfaceData(
-                            id: updated.id,
-                            surfaceType: updated.type,
-                            title: updated.title,
-                            data: updated.data,
-                            actions: updated.actions,
-                            surfaceRef: existing.surfaceRef
-                        )
-                        // Update floating overlay for task_progress cards (macOS only)
-                        #if os(macOS)
-                        if case .card(let cardData) = updated.data,
-                           cardData.template == "task_progress",
-                           let templateData = cardData.templateData,
-                           let progressData = TaskProgressData.parse(from: templateData, fallbackTitle: cardData.title) {
-                            TaskProgressOverlayManager.shared.update(data: progressData, surfaceId: msg.surfaceId)
-                        }
-                        #endif
-                    }
-                    return
-                }
-            }
+            handleSurfaceUpdate(msg)
 
         case .uiSurfaceDismiss(let msg):
-            guard belongsToConversation(msg.conversationId) else { return }
-            // Find and remove the inline surface across all messages
-            for msgIndex in messages.indices {
-                if let surfaceIndex = messages[msgIndex].inlineSurfaces.firstIndex(where: { $0.id == msg.surfaceId }) {
-                    messages[msgIndex].inlineSurfaces.remove(at: surfaceIndex)
-                    return
-                }
-            }
+            handleSurfaceDismiss(msg)
 
         case .uiSurfaceComplete(let msg):
-            guard belongsToConversation(msg.conversationId) else { return }
-            // Dismiss floating overlay for task_progress cards (macOS only)
-            #if os(macOS)
-            TaskProgressOverlayManager.shared.dismiss(surfaceId: msg.surfaceId)
-            #endif
-            // Find the inline surface across all messages and set its completionState
-            for msgIndex in messages.indices {
-                if let surfaceIndex = messages[msgIndex].inlineSurfaces.firstIndex(where: { $0.id == msg.surfaceId }) {
-                    messages[msgIndex].inlineSurfaces[surfaceIndex].completionState = SurfaceCompletionState(
-                        summary: msg.summary,
-                        submittedData: msg.submittedData
-                    )
-                    return
-                }
-            }
+            handleSurfaceComplete(msg)
 
         case .conversationError(let msg):
             // Empty conversationId is treated as a broadcast (e.g. transport-level 401)
@@ -1859,31 +1076,4 @@ extension ChatViewModel {
         }
     }
 
-    /// Auto-open generated video clips in the user's default video player.
-    /// Scans the result for a `clipPath` field rather than checking toolName.
-    /// Restricts to known tool names and validated video extensions to prevent
-    /// arbitrary file opens from untrusted tool results.
-    private static let clipEligibleTools: Set<String> = ["generate_clip"]
-    private static let clipVideoExtensions: Set<String> = ["mp4", "mov", "m4v", "avi", "mkv", "webm"]
-
-    private func autoOpenClipIfNeeded(toolName: String, result: String, isError: Bool) {
-        guard !isError, Self.clipEligibleTools.contains(toolName) else { return }
-        guard let jsonData = result.data(using: .utf8),
-              let json = try? JSONSerialization.jsonObject(with: jsonData) as? [String: Any],
-              let clipPath = json["clipPath"] as? String else {
-            return
-        }
-        let pathExtension = (clipPath as NSString).pathExtension.lowercased()
-        guard Self.clipVideoExtensions.contains(pathExtension) else {
-            log.warning("Clip path has non-video extension '\(pathExtension)', skipping auto-open")
-            return
-        }
-        guard FileManager.default.fileExists(atPath: clipPath) else {
-            log.warning("Clip file not found at path, skipping auto-open")
-            return
-        }
-        #if os(macOS)
-        NSWorkspace.shared.open(URL(fileURLWithPath: clipPath))
-        #endif
-    }
 }
